+++ src/Writers/LaTeXWriter.jl
@@ -27,24 +27,28 @@ makedocs(
 )
 ```
 
-The `makedocs` argument `sitename` will be used for the `\\title` field in the tex document,
-and if the build is for a release tag (i.e. when the `"TRAVIS_TAG"` environment variable is set)
-the version number will be appended to the title.
-The `makedocs` argument `authors` should also be specified, it will be used for the
-`\\authors` field in the tex document.
+The `makedocs` argument `sitename` will be used for the `\\title` field in the tex document.
+The `authors` argument should also be specified and will be used for the `\\authors` field
+in the tex document. Finally, a version number can be specified with the `version` option to
+`LaTeX`, which will be printed in the document and also appended to the output PDF file name.
 
 # Keyword arguments
 
-**`platform`** sets the platform where the tex-file is compiled, either `"native"` (default), `"docker"`,
-or "none" which doesn't compile the tex.
+**`platform`** sets the platform where the tex-file is compiled, either `"native"` (default),
+`"docker"`, or "none" which doesn't compile the tex.
+
+**`version`** specifies the version number that gets printed on the title page of the manual.
+It defaults to the value in the `TRAVIS_TAG` environment variable (although this behaviour is
+considered to be deprecated), or to an empty string if `TRAVIS_TAG` is unset.
 
 See [Other Output Formats](@ref) for more information.
 """
 struct LaTeX <: Documenter.Writer
     platform::String
-    function LaTeX(; platform = "native")
+    version::String
+    function LaTeX(; platform = "native", version = get(ENV, "TRAVIS_TAG", ""))
         platform ∈ ("native", "docker", "none") || throw(ArgumentError("unknown platform: $platform"))
-        return new(platform)
+        return new(platform, string(version))
     end
 end
 
@@ -104,18 +108,16 @@ function render(doc::Documents.Document, settings::LaTeX=LaTeX())
         cp(joinpath(doc.user.root, doc.user.build), joinpath(path, "build"))
         cd(joinpath(path, "build")) do
             name = doc.user.sitename
-            let tag = get(ENV, "TRAVIS_TAG", "")
-                if occursin(Base.VERSION_REGEX, tag)
-                    v = VersionNumber(tag)
-                    name *= "-$(v.major).$(v.minor).$(v.patch)"
-                end
+            if occursin(Base.VERSION_REGEX, settings.version)
+                v = VersionNumber(settings.version)
+                name *= "-$(v.major).$(v.minor).$(v.patch)"
             end
             name = replace(name, " " => "")
             texfile = name * ".tex"
             pdffile = name * ".pdf"
             open(texfile, "w") do io
                 context = Context(io)
-                writeheader(context, doc)
+                writeheader(context, doc, settings)
                 for (title, filename, depth) in files(doc.user.pages)
                     context.filename = filename
                     empty!(context.footnotes)
@@ -212,10 +214,10 @@ function piperun(cmd)
                       stderr = verbose ? stderr : "LaTeXWriter.stderr"))
 end
 
-function writeheader(io::IO, doc::Documents.Document)
+function writeheader(io::IO, doc::Documents.Document, settings::LaTeX)
     custom = joinpath(doc.user.root, doc.user.source, "assets", "custom.sty")
     isfile(custom) ? cp(custom, "custom.sty"; force = true) : touch("custom.sty")
-    
+
     custom_preamble_file = joinpath(doc.user.root, doc.user.source, "assets", "preamble.tex")
     if isfile(custom_preamble_file)
         # copy custom preamble.
@@ -227,11 +229,10 @@ function writeheader(io::IO, doc::Documents.Document)
         """
         % Useful variables
         \\newcommand{\\DocMainTitle}{$(doc.user.sitename)}
-        % Warning: The default value of \\deprecatedEnvTravisTag is deprecated.
-        \\newcommand{\\deprecatedEnvTravisTag}{$(get(ENV, "TRAVIS_TAG", ""))}
+        \\newcommand{\\DocVersion}{$(settings.version)}
         \\newcommand{\\DocAuthors}{$(doc.user.authors)}
         \\newcommand{\\JuliaVersion}{$(VERSION)}
-        
+
         % ---- Insert preamble
         \\input{preamble.tex}
         """